+++ src/uni_modules/wot-design-uni/components/wd-upload/wd-upload.vue
@@ -92,6 +92,8 @@ interface Props {
   // 图片预览相关
   // eslint-disable-next-line @typescript-eslint/ban-types
   beforeChoose?: Function
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  buildFormData?: Function
 }
 
 const props = withDefaults(defineProps<Props>(), {
@@ -147,7 +149,7 @@ watch(
   () => props.limit,
   (val) => {
     if (val && val < uploadFiles.value.length) {
-      throw Error('[wot-design]Error: props limit must less than fileList.length')
+      console.error('[wot-design]Error: props limit must less than fileList.length')
     }
   },
   {
@@ -159,8 +161,8 @@ watch(
 watch(
   () => props.beforePreview,
   (fn) => {
-    if (fn && getType(fn) !== 'function') {
-      throw Error('The type of beforePreview must be Function')
+    if (fn && getType(fn) !== 'function' && getType(fn) !== 'asyncfunction') {
+      console.error('The type of beforePreview must be Function')
     }
   },
   {
@@ -172,8 +174,8 @@ watch(
 watch(
   () => props.onPreviewFail,
   (fn) => {
-    if (fn && getType(fn) !== 'function') {
-      throw Error('The type of onPreviewFail must be Function')
+    if (fn && getType(fn) !== 'function' && getType(fn) !== 'asyncfunction') {
+      console.error('The type of onPreviewFail must be Function')
     }
   },
   {
@@ -185,8 +187,8 @@ watch(
 watch(
   () => props.beforeRemove,
   (fn) => {
-    if (fn && getType(fn) !== 'function') {
-      throw Error('The type of beforeRemove must be Function')
+    if (fn && getType(fn) !== 'function' && getType(fn) !== 'asyncfunction') {
+      console.error('The type of beforeRemove must be Function')
     }
   },
   {
@@ -198,8 +200,8 @@ watch(
 watch(
   () => props.beforeUpload,
   (fn) => {
-    if (fn && getType(fn) !== 'function') {
-      throw Error('The type of beforeUpload must be Function')
+    if (fn && getType(fn) !== 'function' && getType(fn) !== 'asyncfunction') {
+      console.error('The type of beforeUpload must be Function')
     }
   },
   {
@@ -211,8 +213,21 @@ watch(
 watch(
   () => props.beforeChoose,
   (fn) => {
-    if (fn && getType(fn) !== 'function') {
-      throw Error('The type of beforeChoose must be Function')
+    if (fn && getType(fn) !== 'function' && getType(fn) !== 'asyncfunction') {
+      console.error('The type of beforeChoose must be Function')
+    }
+  },
+  {
+    deep: true,
+    immediate: true
+  }
+)
+
+watch(
+  () => props.buildFormData,
+  (fn) => {
+    if (fn && getType(fn) !== 'function' && getType(fn) !== 'asyncfunction') {
+      console.error('The type of buildFormData must be Function')
     }
   },
   {
@@ -231,6 +246,8 @@ function initFile(file) {
   // 状态初始化
   const initState = {
     uid: context.id++,
+    // 仅h5支持 name
+    name: file.name || '',
     status: 'loading',
     size: file.size,
     url: file.path,
@@ -239,7 +256,20 @@ function initFile(file) {
   }
 
   uploadFiles.value.push(initState)
-  handleUpload(initState)
+
+  const { buildFormData, formData = {} } = props
+
+  if (buildFormData) {
+    buildFormData({
+      file: initState,
+      formData,
+      resolve: (formData: Record<string, any>) => {
+        formData && handleUpload(initState, formData)
+      }
+    })
+  } else {
+    handleUpload(initState, formData)
+  }
 }
 
 /**
@@ -247,14 +277,14 @@ function initFile(file) {
  * @param {Object} err 错误返回信息
  * @param {Object} file 上传的文件
  */
-function handleError(err, file) {
+function handleError(err, file, formData: Record<string, any>) {
   const { statusKey } = props
   const index = uploadFiles.value.findIndex((item) => item.uid === file.uid)
   if (index > -1) {
     uploadFiles.value[index][statusKey] = 'fail'
     uploadFiles.value[index].error = err.message
     uploadFiles.value[index].response = err
-    emit('fail', { error: err, file })
+    emit('fail', { error: err, file, formData })
   }
 }
 
@@ -263,14 +293,14 @@ function handleError(err, file) {
  * @param {Object} res 接口返回信息
  * @param {Object} file 上传的文件
  */
-function handleSuccess(res, file) {
+function handleSuccess(res, file, formData: Record<string, any>) {
   const { statusKey } = props
   const index = uploadFiles.value.findIndex((item) => item.uid === file.uid)
   if (index > -1) {
     uploadFiles.value[index][statusKey] = 'success'
     uploadFiles.value[index].response = res.data
     emit('change', { fileList: uploadFiles.value })
-    emit('success', { file, fileList: uploadFiles.value })
+    emit('success', { file, fileList: uploadFiles.value, formData })
   }
 }
 
@@ -291,8 +321,9 @@ function handleProgress(res, file) {
  * @description 上传操作
  * @param {Object} file 上传的文件
  */
-function handleUpload(file) {
-  const { action, name, formData = {}, header = {}, accept } = props
+function handleUpload(file, formData: Record<string, any>) {
+  const { action, name, header = {}, accept } = props
+
   const uploadTask = uni.uploadFile({
     url: action,
     header,
@@ -302,12 +333,17 @@ function handleUpload(file) {
     formData,
     filePath: file.url,
     success(res) {
-      // 上传成功进行文件列表拼接
-      handleSuccess(res, file)
+      if (res.statusCode === 200) {
+        // 上传成功进行文件列表拼接
+        handleSuccess(res, file, formData)
+      } else {
+        // 上传失败处理
+        handleError(res, file, formData)
+      }
     },
     fail(err) {
       // 上传失败处理
-      handleError(err, file)
+      handleError(err, file, formData)
     }
   })