+++ src/uni_modules/wot-design-uni/global.d.ts
@@ -1,7 +1,7 @@
 /*
  * @Author: weisheng
  * @Date: 2023-09-25 17:28:12
- * @LastEditTime: 2023-11-03 23:21:14
+ * @LastEditTime: 2023-12-06 16:42:32
  * @LastEditors: weisheng
  * @Description:
  * @FilePath: \wot-design-uni\src\uni_modules\wot-design-uni\global.d.ts
@@ -89,6 +89,10 @@ declare module '@vue/runtime-core' {
     WdTableCol: typeof import('./components/wd-table-col/wd-table-col.vue')['default']
     WdSidebar: typeof import('./components/wd-sidebar/wd-sidebar.vue')['default']
     WdSidebarItem: typeof import('./components/wd-sidebar-item/wd-sidebar-item.vue')['default']
+    WdFab: typeof import('./components/wd-fab/wd-fab.vue')['default']
+    WdCountDown: typeof import('./components/wd-count-down/wd-count-down.vue')['default']
+    WdNumberKeyboard: typeof import('./components/wd-number-keyboard/wd-number-keyboard.vue')['default']
+    WdGap: typeof import('./components/wd-gap/wd-gap.vue')['default']
   }
 }